updates to style and organization
